feat : Post 데이터 클래스 수정 #76
